// src/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
      places: {
        Row: {
          author: string | null
          category: Database["public"]["Enums"]["place_category"]
          created_at: string
          description: string
          id: number
          lat: number
          lng: number
          name: string
          status: Database["public"]["Enums"]["place_status"]
          terrain_types: Database["public"]["Enums"]["terrain_type"][]
        }
        Insert: {
          author?: string | null
          category?: Database["public"]["Enums"]["place_category"]
          created_at?: string
          description?: string
          id?: number
          lat: number
          lng: number
          name?: string
          status?: Database["public"]["Enums"]["place_status"]
          terrain_types?: Database["public"]["Enums"]["terrain_type"][]
        }
        Update: {
          author?: string | null
          category?: Database["public"]["Enums"]["place_category"]
          created_at?: string
          description?: string
          id?: number
          lat?: number
          lng?: number
          name?: string
          status?: Database["public"]["Enums"]["place_status"]
          terrain_types?: Database["public"]["Enums"]["terrain_type"][]
        }
        Relationships: [
          {
            foreignKeyName: "places_author_fkey1"
            columns: ["author"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      profiles: {
        Row: {
          avatar_url: string | null
          created_at: string | null
          id: string
          privacy_policy_accepted_at: string | null
          privacy_policy_version: string | null
          role: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          created_at?: string | null
          id: string
          privacy_policy_accepted_at?: string | null
          privacy_policy_version?: string | null
          role?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          created_at?: string | null
          id?: string
          privacy_policy_accepted_at?: string | null
          privacy_policy_version?: string | null
          role?: string | null
          username?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      generate_username: { Args: never; Returns: string }
    }
    Enums: {
      place_category: "legal" | "semi-legal" | "closed"
      place_status: "pending" | "declined" | "active"
      terrain_type: "sand" | "mud" | "forest" | "gravel" | "rocks" | "asphalt"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never